% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_checkr.R
\name{dice_checkr}
\alias{dice_checkr}
\title{The dice_checkr function is designed to simulate rolling a dice a number of
and obtain a value of better than based on the number of sides on the dice}
\usage{
dice_checkr(
  sides = 1:6,
  rolls = 1000,
  value = 3,
  rerolls = TRUE,
  reroll_val = 1:2,
  params = ...
)
}
\arguments{
\item{sides}{the number of faces the dice has}

\item{rolls}{the number of rolls you wish to simulate}

\item{value}{the value or better you are wishing to obtain}

\item{rerolls}{can you reroll the dice}

\item{reroll_val}{are there any values in particular to reroll, just a roll of
1 for examples?}

\item{params}{holder argument, not currently used in this function}
}
\value{
a scalar with proportion of successful rolls,
will be equivalent to the general probably i.e. 1/6 if looking for a 6 on a
standard die
}
\description{
The dice_checkr function is designed to simulate rolling a dice a number of
and obtain a value of better than based on the number of sides on the dice
}
\examples{
dice_checkr(value = 6, rerolls = F)

}
