% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_distro.R
\name{dice_distro}
\alias{dice_distro}
\title{Distribution for desired values when for rolling x number of dice y times
The aim is to simulate the expected variance}
\usage{
dice_distro(
  sides = 1:6,
  rolls = 1000,
  value = 3,
  times = 1:1000,
  rerolls = FALSE,
  reroll_val = 1:2
)
}
\arguments{
\item{sides}{Number of sides or faces on the die}

\item{rolls}{the number of rolls you wish to make}

\item{value}{the value or better you wish to obtain}

\item{times}{the number of time you wish to roll the dice}

\item{rerolls}{can you reroll certain value}

\item{reroll_val}{the value to be rerolled defaults to 1:2}
}
\value{
A list with the total successes in a per times rolled
The average probability of success
The max probability of success across all rolls
The min probability of success across all rolls
The upper and lower bound of probability to approximately 99\% of a normal
distribution.
The average number of successes to expectbased on the simulation.
}
\description{
Distribution for desired values when for rolling x number of dice y times
The aim is to simulate the expected variance
}
\examples{
# roll a 2 or better on 20 dice
dice_distro(rolls = 20, value = 2, rerolls = F)
}
