% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wound_checker.R
\name{wound_checker}
\alias{wound_checker}
\title{Provide the number of successful wound expected based on the proportion of
likely hits, strength and toughness characteristic of the attack}
\usage{
wound_checker(
  dice = 12,
  hit_prob = 0.66,
  strength = 4,
  toughness = 4,
  lethal_hits = FALSE,
  reroll_wounds = FALSE
)
}
\arguments{
\item{dice}{number of dice rolled}

\item{hit_prob}{probability of a successful hit}

\item{strength}{strength of attacking model}

\item{toughness}{toughness of defending model}

\item{lethal_hits}{are the wound lethal? Set to FALSE}

\item{reroll_wounds}{are wounds rerolled as part of the attack? Set to FALSE}
}
\value{
a numeric scalar object
}
\description{
Numbers are rounded up to the nearest whole number
}
\examples{
wound_checker(dice = 24, hit_prob = 0.33, strength = 4, toughness = 5)
}
