% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wound_probability.R
\name{wound_probability}
\alias{wound_probability}
\title{Find the probability of wounding based on strength and wound characteristics}
\usage{
wound_probability(strength = 4, toughness = 4)
}
\arguments{
\item{strength}{the strength of the attacking model}

\item{toughness}{the toughness of the defending model}
}
\value{
the likely success based on the probability to 2 decimal places
}
\description{
Find the probability of wounding based on strength and wound characteristics
}
\examples{
wound_probability(strength = 4, toughness = 5)
}
