% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wound_table.R
\name{wound_table}
\alias{wound_table}
\title{Title}
\usage{
wound_table(
  dice = 24,
  strength = c(2, 3, 4, 5, 8),
  toughess = 4,
  lethals = FALSE,
  hit_prob = 0.66
)
}
\arguments{
\item{dice}{the number of dice to roll}

\item{strength}{the strength of the user}

\item{toughess}{the toughness}

\item{lethals}{whether lethal are set}

\item{hit_prob}{the probability of hit}
}
\value{
a data frame
}
\description{
Title
}
